COMMERCE-1276 Only try to delete workflowInstanceLinks if they exist

+++ commerce-service/src/main/java/com/liferay/commerce/service/impl/CommerceOrderLocalServiceImpl.java
@@ -398,10 +398,16 @@ public class CommerceOrderLocalServiceImpl
 
 		CommerceAccount commerceAccount = commerceOrder.getCommerceAccount();
 
-		workflowInstanceLinkLocalService.deleteWorkflowInstanceLinks(
-			commerceOrder.getCompanyId(),
-			commerceAccount.getCommerceAccountGroupId(),
-			CommerceOrder.class.getName(), commerceOrder.getCommerceOrderId());
+		if (hasWorkflowDefinition(
+				commerceOrder.getScopeGroupId(),
+				commerceOrder.getCommerceOrderId())) {
+
+			workflowInstanceLinkLocalService.deleteWorkflowInstanceLinks(
+				commerceOrder.getCompanyId(),
+				commerceAccount.getCommerceAccountGroupId(),
+				CommerceOrder.class.getName(),
+				commerceOrder.getCommerceOrderId());
+		}
 
 		return commerceOrder;
 	}
